Add cairo_run task and start fresh run_pipeline.py
